@@ -20,88 +20,88 @@
     "test": "react-app-rewired test --passWithNoTests"
   },
   "dependencies": {
-    "@lingui/react": "^3.8.10",
-    "@sentry/browser": "^5.27.6",
-    "@types/bcryptjs": "^2.4.2",
-    "@types/papaparse": "^5.2.5",
-    "@types/validator": "^13.0.0",
-    "axios": "^0.21.4",
-    "boxicons": "^2.0.5",
-    "buffer": "^5.7.1",
-    "classnames": "^2.2.6",
-    "downloadjs": "^1.4.7",
-    "draft-js": "^0.11.7",
-    "draftjs-to-html": "^0.9.1",
-    "escape-html": "^1.0.3",
-    "html-to-draftjs": "^1.5.0",
-    "lodash": "^4.17.21",
-    "lottie-react": "^2.1.0",
-    "moment": "^2.29.1",
-    "papaparse": "^5.2.0",
-    "querystring": "^0.2.1",
-    "react": "^18.0.0",
-    "react-app-polyfill": "^3.0.0",
-    "react-dom": "^18.0.0",
-    "react-draft-wysiwyg": "^1.14.6",
-    "react-ga": "^3.3.0",
-    "react-moment": "^1.1.1",
-    "react-paginate": "^8.1.1",
-    "react-player": "^2.6.2",
-    "react-router-dom": "^5.1.2",
-    "react-textarea-autosize": "^8.3.3",
-    "sass-burger": "^1.3.1",
-    "sms-segments-calculator": "^1.1.1",
-    "spark-md5": "^3.0.1",
-    "stream": "^0.0.2",
-    "typescript": "^4.2.3",
-    "url": "^0.11.0",
-    "util": "^0.12.4",
-    "uuidv4": "^6.1.1",
-    "validator": "^13.7.0",
-    "webcrypto-shim": "^0.1.5"
+    "@lingui/react": "3.13.3",
+    "@sentry/browser": "5.27.6",
+    "@types/bcryptjs": "2.4.2",
+    "@types/papaparse": "5.2.5",
+    "@types/validator": "13.0.0",
+    "axios": "0.21.4",
+    "boxicons": "2.0.5",
+    "buffer": "5.7.1",
+    "classnames": "2.2.6",
+    "downloadjs": "1.4.7",
+    "draft-js": "0.11.7",
+    "draftjs-to-html": "0.9.1",
+    "escape-html": "1.0.3",
+    "html-to-draftjs": "1.5.0",
+    "lodash": "4.17.21",
+    "lottie-react": "2.2.1",
+    "moment": "2.29.3",
+    "papaparse": "5.2.0",
+    "querystring": "0.2.1",
+    "react": "18.1.0",
+    "react-app-polyfill": "3.0.0",
+    "react-dom": "18.1.0",
+    "react-draft-wysiwyg": "1.14.6",
+    "react-ga": "3.3.0",
+    "react-moment": "1.1.1",
+    "react-paginate": "8.1.1",
+    "react-player": "2.6.2",
+    "react-router-dom": "5.1.2",
+    "react-textarea-autosize": "8.3.3",
+    "sass-burger": "1.3.1",
+    "sms-segments-calculator": "1.1.1",
+    "spark-md5": "3.0.1",
+    "stream": "0.0.2",
+    "typescript": "4.2.3",
+    "url": "0.11.0",
+    "util": "0.12.4",
+    "uuidv4": "6.1.1",
+    "validator": "13.7.0",
+    "webcrypto-shim": "0.1.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.4",
-    "@lingui/cli": "^3.5.1",
-    "@lingui/macro": "^3.5.1",
-    "@peculiar/webcrypto": "^1.1.6",
-    "@sentry/cli": "^1.64.2",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.1",
-    "@testing-library/user-event": "^14.1.0",
-    "@types/classnames": "^2.2.10",
-    "@types/downloadjs": "^1.4.2",
-    "@types/draft-js": "^0.10.44",
-    "@types/draftjs-to-html": "^0.8.0",
+    "@babel/core": "7.18.2",
+    "@lingui/cli": "3.5.1",
+    "@lingui/macro": "3.5.1",
+    "@peculiar/webcrypto": "1.1.6",
+    "@sentry/cli": "1.64.2",
+    "@testing-library/jest-dom": "5.16.4",
+    "@testing-library/react": "13.3.0",
+    "@testing-library/user-event": "14.1.0",
+    "@types/classnames": "2.2.10",
+    "@types/downloadjs": "1.4.2",
+    "@types/draft-js": "0.10.44",
+    "@types/draftjs-to-html": "0.8.0",
     "@types/escape-html": "0.0.20",
-    "@types/html-to-draftjs": "^1.4.0",
-    "@types/jest": "^26.0.21",
-    "@types/lodash": "^4.14.165",
-    "@types/node": "^12.12.34",
-    "@types/react": "^16.14.5",
-    "@types/react-dom": "^16.9.12",
-    "@types/react-draft-wysiwyg": "^1.13.0",
-    "@types/react-paginate": "^6.2.1",
-    "@types/react-router-dom": "^5.1.3",
-    "@types/react-textarea-autosize": "^4.3.5",
-    "@types/spark-md5": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.11.0",
-    "@typescript-eslint/parser": "^5.11.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "eslint": "^8.8.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "jest-canvas-mock": "^2.3.1",
-    "lint-staged": "^10.2.6",
-    "msw": "^0.39.2",
-    "prettier": "^2.0.5",
-    "react-app-rewire-alias": "^1.1.7",
-    "react-app-rewired": "^2.1.11",
-    "react-scripts": "^5.0.0",
-    "sass": "^1.49.7"
+    "@types/html-to-draftjs": "1.4.0",
+    "@types/jest": "26.0.21",
+    "@types/lodash": "4.14.165",
+    "@types/node": "12.12.34",
+    "@types/react": "16.14.5",
+    "@types/react-dom": "16.9.12",
+    "@types/react-draft-wysiwyg": "1.13.0",
+    "@types/react-paginate": "6.2.1",
+    "@types/react-router-dom": "5.1.3",
+    "@types/react-textarea-autosize": "4.3.5",
+    "@types/spark-md5": "3.0.2",
+    "@typescript-eslint/eslint-plugin": "5.11.0",
+    "@typescript-eslint/parser": "5.11.0",
+    "babel-core": "7.0.0-bridge.0",
+    "eslint": "8.8.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-import-resolver-typescript": "2.5.0",
+    "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-react": "7.28.0",
+    "jest-canvas-mock": "2.3.1",
+    "lint-staged": "10.2.6",
+    "msw": "0.39.2",
+    "prettier": "2.0.5",
+    "react-app-rewire-alias": "1.1.7",
+    "react-app-rewired": "2.1.11",
+    "react-scripts": "5.0.0",
+    "sass": "1.49.7"
   },
   "eslintConfig": {
     "extends": [
