@@ -1,15 +1,6 @@
 import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
 import config from '../core/config'
 
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace NodeJS {
-    interface Global {
-      sequelize: Sequelize
-    }
-  }
-}
-
 // fix PGUSER and PGPASSWORD inside this global test setup to dictate test db
 // credentials for surrounding setup like local docker compose or CI containers
 process.env.PGUSER = 'postgres'
@@ -20,15 +11,15 @@ const TEST_DB = 'postmangovsg_test'
 // defined in  npm test command
 const JEST_WORKERS = 2
 
-module.exports = async (): Promise<void> => {
-  global.sequelize = new Sequelize(DB_URI, {
-    dialect: 'postgres',
-    logging: false,
-    pool: config.get('database.poolOptions'),
-  } as SequelizeOptions)
+export const sequelize = new Sequelize(DB_URI, {
+  dialect: 'postgres',
+  logging: false,
+  pool: config.get('database.poolOptions'),
+} as SequelizeOptions)
 
+export default async (): Promise<void> => {
   for (let i = 1; i <= JEST_WORKERS; i++) {
-    await global.sequelize.query(`DROP DATABASE IF EXISTS ${TEST_DB}_${i}`)
-    await global.sequelize.query(`CREATE DATABASE ${TEST_DB}_${i}`)
+    await sequelize.query(`DROP DATABASE IF EXISTS ${TEST_DB}_${i}`)
+    await sequelize.query(`CREATE DATABASE ${TEST_DB}_${i}`)
   }
 }
