@@ -1,3 +1,5 @@
+import { sequelize } from './global-setup'
+
 module.exports = async function (): Promise<void> {
-  await global.sequelize.close()
+  await sequelize.close()
 }
